refactor. 출고 기능 리팩토링
